requirements.txt
duckdb
pyarrow
pandas
requests
python-dotenv